web/web/requirements.txt
Flask
python-dotenv
supabase
pytz
